+++ projectroles/tests/test_views_api.py
@@ -50,6 +50,10 @@ from projectroles.tests.test_views import (
     REMOTE_SITE_SECRET,
 )
 from projectroles.utils import build_secret
+from projectroles.views_api import (
+    SODAR_API_MEDIA_TYPE,  # TODO: Remove in v1.1 (see #1401)
+    SODAR_API_DEFAULT_VERSION,  # TODO: Remove in v1.1 (see #1401)
+)
 
 
 app_settings = AppSettingAPI()
@@ -129,8 +133,8 @@ class SODARAPIViewTestMixin(SerializedObjectMixin):
     # Default API header parameters are for external SODAR site APIs
     # DEPRECATED: To be removed in SODAR Core v1.1 (see #1401)
     # Instead, provide a media type and version specific to your app
-    media_type = settings.SODAR_API_MEDIA_TYPE
-    api_version = settings.SODAR_API_DEFAULT_VERSION
+    media_type = SODAR_API_MEDIA_TYPE
+    api_version = SODAR_API_DEFAULT_VERSION
 
     # Copied from Knox tests
     @classmethod